+++ geotrek/api/management/commands/sync_mobile.py
@@ -1,16 +1,15 @@
 import argparse
-import logging
 import filecmp
+import logging
 import os
-import stat
-from PIL import Image
 import re
 import shutil
+import stat
 import tempfile
 from time import sleep
 from zipfile import ZipFile
-import cairosvg
 
+import cairosvg
 from django.conf import settings
 from django.contrib.auth.models import AnonymousUser
 from django.core.management.base import BaseCommand, CommandError
@@ -19,19 +18,21 @@ from django.http import StreamingHttpResponse
 from django.test.client import RequestFactory
 from django.utils import translation
 from django.utils.translation import gettext as _
-from geotrek.common.models import FileType  # NOQA
+from modeltranslation.utils import build_localized_fieldname
+from PIL import Image
+
+from geotrek.api.mobile.views.common import FlatPageViewSet, SettingsView
+from geotrek.api.mobile.views.trekking import TrekViewSet
 from geotrek.common import models as common_models
 from geotrek.common.functions import GeometryType
+from geotrek.common.helpers_sync import ZipTilesBuilder
+from geotrek.common.models import FileType  # NOQA
 from geotrek.flatpages.models import FlatPage
 from geotrek.tourism import models as tourism_models
-from geotrek.trekking import models as trekking_models
-from geotrek.api.mobile.views.trekking import TrekViewSet
-from geotrek.api.mobile.views.common import FlatPageViewSet, SettingsView
-from geotrek.common.helpers_sync import ZipTilesBuilder
-# Register mapentity models
-from geotrek.trekking import urls  # NOQA
 from geotrek.tourism import urls  # NOQA
-
+# Register mapentity models
+from geotrek.trekking import models as trekking_models
+from geotrek.trekking import urls  # NOQA
 
 logger = logging.getLogger(__name__)
 
@@ -237,7 +238,7 @@ class Command(BaseCommand):
 
     def sync_flatpage(self, lang):
         flatpages = FlatPage.objects.order_by('pk').filter(target__in=['mobile', 'all']).filter(
-            **{'published_{lang}'.format(lang=lang): True})
+            **{build_localized_fieldname('published', lang): True})
         if self.portal:
             flatpages = flatpages.filter(Q(portal__name__in=self.portal) | Q(portal=None))
         self.sync_json(lang, FlatPageViewSet, 'flatpages',
@@ -249,7 +250,7 @@ class Command(BaseCommand):
     def sync_trekking(self, lang):
         self.sync_geojson(lang, TrekViewSet, 'treks.geojson', type_view={'get': 'list'})
         treks = trekking_models.Trek.objects.annotate(geom_type=GeometryType("geom")).filter(geom_type="LINESTRING").existing().order_by('pk')
-        treks = treks.filter(**{'published_{lang}'.format(lang=lang): True})
+        treks = treks.filter(**{build_localized_fieldname('published', lang): True})
 
         if self.portal:
             treks = treks.filter(Q(portal__name__in=self.portal) | Q(portal=None))