+++ geotrek/common/forms.py
@@ -18,6 +18,7 @@ from django.urls import reverse
 from django.utils.text import format_lazy
 from django.utils.translation import gettext_lazy as _
 from mapentity.forms import MapEntityForm, SubmitButton
+from modeltranslation.utils import build_localized_fieldname
 
 from geotrek.authent.models import (StructureOrNoneRelated, StructureRelated,
                                     default_structure)
@@ -196,7 +197,7 @@ class CommonForm(MapEntityForm):
     @property
     def any_published(self):
         """Check if form has published in at least one of the language"""
-        return any([self.cleaned_data.get(f'published_{language[0]}', False)
+        return any([self.cleaned_data.get(build_localized_fieldname('published', language[0]), False)
                     for language in settings.MAPENTITY_CONFIG['TRANSLATED_LANGUAGES']])
 
     @property
@@ -205,7 +206,7 @@ class CommonForm(MapEntityForm):
         """
         languages = [language[0] for language in settings.MAPENTITY_CONFIG['TRANSLATED_LANGUAGES']]
         if settings.PUBLISHED_BY_LANG:
-            return [language for language in languages if self.cleaned_data.get(f'published_{language}', None)]
+            return [language for language in languages if self.cleaned_data.get(build_localized_fieldname('published', language), None)]
         else:
             if self.any_published:
                 return languages
@@ -237,12 +238,12 @@ class CommonForm(MapEntityForm):
             if field_required in translated_fields:
                 if self.cleaned_data.get('review') and settings.COMPLETENESS_LEVEL == 'error_on_review':
                     # get field for first language only
-                    field_required_lang = f"{field_required}_{settings.MAPENTITY_CONFIG['TRANSLATED_LANGUAGES'][0][0]}"
+                    field_required_lang = build_localized_fieldname(field_required, settings.MAPENTITY_CONFIG['TRANSLATED_LANGUAGES'][0][0])
                     missing_fields.append(field_required_lang)
                     self.add_error(field_required_lang, msg)
                 else:
                     for language in self.published_languages:
-                        field_required_lang = f'{field_required}_{language}'
+                        field_required_lang = build_localized_fieldname(field_required, language)
                         if not self.cleaned_data.get(field_required_lang):
                             missing_fields.append(field_required_lang)
                             self.add_error(field_required_lang, msg)