+++ docs/changelog.rst
@@ -34,6 +34,11 @@ CHANGELOG
 - Extract all geometry types in views `v_outdoor_sites` and `v_outdoor_courses` (#3603)
 - Maintenance appears several times on some zoning filters (#3881)
 
+**Bug fixes**
+
+- Fix sub-language usage (en-US, zh-hant, ...) (#3801)
+
+
 2.101.4     (2023-11-15)
 ------------------------